The five codex-pipeline cycles are gone, and the four that remain cannot fire
Each `test_codex_pipeline_part*` imported three test doubles back from the parent, which imports
each part's mixin class. Every one of the five failed when reached first:

    ImportError: cannot import name '_CodexPipelinePart1' from partially initialized module

matrixarkai#1597 and matrixarkai#1603 removed the same cycle from two other clusters by importing
directly, because those borrowed nothing but stdlib and production modules. This one shares real
test doubles -- CountingLocalAdapter, FastHookLocalAdapter, NativeCaptureLocalAdapter, 62 lines --
so they move to `test_codex_pipeline_fixtures`, which imports no test module. The parent and all
five parts read from there, and the graph is acyclic:

    each part imported first    ok, all five
    the parent                  44 failures, 10 errors -- identical to main, the same ones

The fixtures module copies the parent's import block VERBATIM, without a `tools.` prefix, because
that prefix would make a different module object with its own state and a test patching one through
the parent would not reach a part holding the other. That cost four tests in matrixarkai#1603.

They stay in KNOWN: importing the fixtures module is still a cross-test import. What they are no
longer is MUTUAL, which is the half that reorders discover and fails tests they have nothing to do
with.

  KNOWN_MUTUAL  9 -> 4

AND THE FOUR THAT REMAIN CANNOT FIRE

I was about to move 425 lines of backend-policy fixtures before checking. Those four parts never
reach their cycle: `test_matrixark_mcp_backend_policy` raises SkipTest at import because
`run_matrixark_rust_scale_report` is absent from this repository and deliberately not recreated --
the symbols it owes name deployment infrastructure and a stand-in would be guessing at scrubbed
content. Their tests cannot run here at all, so breaking the cycle would have changed nothing and
shown up as a pair count falling 9 to 5.

That is recorded beside the list, along with two things the sizing turned up that would have made
the move wrong anyway: `_FailingWarmupClient` is used by another fixture but borrowed by no
part, and `mcp_core` is not among the parent's module-level imports.
