+++ tools/test_codex_pipeline_part3.py
@@ -12,7 +12,7 @@ except ImportError:
     from matrixark_mcp_core import *  # noqa: F401,F403
 
 try:  # names owned by the parent module
-    from tools.test_matrixark_codex_hook_pipeline import (
+    from test_codex_pipeline_fixtures import (
     FastHookLocalAdapter,
     MatrixArkLocalAdapter,
     Namespace,
@@ -28,7 +28,7 @@ try:  # names owned by the parent module
     time,
 )
 except ImportError:
-    from test_matrixark_codex_hook_pipeline import (
+    from test_codex_pipeline_fixtures import (
     FastHookLocalAdapter,
     MatrixArkLocalAdapter,
     Namespace,
